requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
docopt==0.6.2
frozenlist==1.3.3
idna==3.4
Js2Py==0.74
multidict==6.0.4
openai==0.27.5
packaging==23.1
pipwin==0.5.2
PyAudio==0.2.13
pycodestyle==2.10.0
pycparser==2.21
pyjsparser==2.7.1
PyPrind==2.11.3
pySmartDL==1.3.4
python-dotenv==1.0.0
pyttsx3==2.90
pytz-deprecation-shim==0.1.0.post0
requests==2.29.0
six==1.16.0
sounddevice==0.4.6
soupsieve==2.4.1
SpeechRecognition==3.10.0
tomli==2.0.1
tqdm==4.65.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
yarl==1.9.2
